+++ documentation/changelog.md
@@ -38,6 +38,7 @@ description: List of all the changes and current versions.
 + Fix Hologram Invincibility
 + Add Custom Vehicle Rules for engine, unlock, lock, doors, enter, and exit
 + Add Custom Storage Rules for opening
++ Add Custom Interior Rules for enter, exit, unlock, and lock
 ```
 
 ## 2.0.3